+++ examples/user-managed-google-managed-ssl/main.tf
@@ -65,7 +65,7 @@ locals {
 
 module "gce-lb-https" {
   source  = "terraform-google-modules/lb-http/google"
-  version = "~> 10.0"
+  version = "~> 12.0"
 
   name                            = var.network_name
   project                         = var.project_id
@@ -99,7 +99,7 @@ module "gce-lb-https" {
       ]
 
       iap_config = {
-        enable = false
+        enable = true
       }
     }
 
@@ -122,7 +122,7 @@ module "gce-lb-https" {
       ]
 
       iap_config = {
-        enable = false
+        enable = true
       }
     }
   }