+++ examples/user-managed-google-managed-ssl/mig.tf
@@ -24,7 +24,7 @@ data "template_file" "group1-startup-script" {
 
 module "mig1_template" {
   source     = "terraform-google-modules/vm/google//modules/instance_template"
-  version    = "~> 11.1"
+  version    = "~> 12.0"
   network    = google_compute_network.default.self_link
   subnetwork = google_compute_subnetwork.group1.self_link
   service_account = {
@@ -33,7 +33,7 @@ module "mig1_template" {
   }
   name_prefix          = "${var.network_name}-group1"
   startup_script       = data.template_file.group1-startup-script.rendered
-  source_image_family  = "ubuntu-2004-lts"
+  source_image_family  = "ubuntu-2204-lts"
   source_image_project = "ubuntu-os-cloud"
   tags = [
     "${var.network_name}-group1",
@@ -43,7 +43,7 @@ module "mig1_template" {
 
 module "mig1" {
   source            = "terraform-google-modules/vm/google//modules/mig"
-  version           = "~> 11.1"
+  version           = "~> 12.0"
   instance_template = module.mig1_template.self_link
   region            = var.group1_region
   hostname          = "${var.network_name}-group1"